Added CLI
